+++ src/main/java/de/muenchen/dave/domain/dtos/laden/LadeBelastungsplanFjsDTO.java
@@ -1,0 +1,10 @@
+package de.muenchen.dave.domain.dtos.laden;
+
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+
+@Data
+@EqualsAndHashCode(callSuper = true)
+public class LadeBelastungsplanFjsDTO extends AbstractLadeBelastungsplanDTO<BelastungsplanFjsDataDTO> {
+
+}
