+++ src/main/java/de/muenchen/dave/services/processzaehldaten/ProcessZaehldatenBelastungsplanService.java
@@ -7,8 +7,10 @@ import de.muenchen.dave.domain.dtos.OptionsDTO;
 import de.muenchen.dave.domain.dtos.laden.AbstractBelastungsplanDataDTO;
 import de.muenchen.dave.domain.dtos.laden.AbstractLadeBelastungsplanDTO;
 import de.muenchen.dave.domain.dtos.laden.BelastungsplanDataDTO;
+import de.muenchen.dave.domain.dtos.laden.BelastungsplanFjsDataDTO;
 import de.muenchen.dave.domain.dtos.laden.BelastungsplanQjsDataDTO;
 import de.muenchen.dave.domain.dtos.laden.LadeBelastungsplanDTO;
+import de.muenchen.dave.domain.dtos.laden.LadeBelastungsplanFjsDTO;
 import de.muenchen.dave.domain.dtos.laden.LadeBelastungsplanQjsDTO;
 import de.muenchen.dave.domain.dtos.laden.LadeZaehldatumDTO;
 import de.muenchen.dave.domain.dtos.laden.LadeZaehldatumTageswertDTO;
@@ -45,6 +47,7 @@ import java.util.stream.Stream;
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 import lombok.extern.slf4j.Slf4j;
@@ -55,6 +58,7 @@ import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 
 @Service
+@RequiredArgsConstructor
 @Slf4j
 public class ProcessZaehldatenBelastungsplanService {
 
@@ -66,13 +70,7 @@ public class ProcessZaehldatenBelastungsplanService {
 
     private final LadeZaehldatenService ladeZaehldatenService;
 
-    public ProcessZaehldatenBelastungsplanService(final ZeitintervallRepository zeitintervallRepository,
-            final ZaehlstelleIndex zaehlstelleIndex,
-            final LadeZaehldatenService ladeZaehldatenService) {
-        this.zeitintervallRepository = zeitintervallRepository;
-        this.zaehlstelleIndex = zaehlstelleIndex;
-        this.ladeZaehldatenService = ladeZaehldatenService;
-    }
+    private final BelastungsplanDataServiceFactory belastungsplanDataServiceFactory;
 
     /**
      * Subtrahiert alle BigDecimal[][] des vergleichsBelastungsplans von den BigDecimal[][] des
@@ -160,48 +158,6 @@ public class ProcessZaehldatenBelastungsplanService {
             differences[i] = basis[i].subtract(vergleich[i]);
         }
         return differences;
-    }
-
-    private static boolean isKreisverkehr(final Verkehrsbeziehung verkehrsbeziehung) {
-        return ObjectUtils.isNotEmpty(verkehrsbeziehung.getFahrbewegungKreisverkehr())
-                && ObjectUtils.isEmpty(verkehrsbeziehung.getNach());
-    }
-
-    /**
-     * Die Grafiken im Frontend erwarten pro Verkehrsbeziehung einen einzelnen Wert. Um an alle Werte
-     * mittels Index zugreifen zu können ist ein 2-Stufiges Array
-     * erforderlich. Ebene 1: Enthält alle Werte für die Von-Spuren Ebene 2: Enthält die Werte für die
-     * Nach-Spur pro Von-Spur Bsp.: [ [Nach_1, Nach_2, ...,
-     * Nach_8] // Von_1 ... [Nach_1, Nach_2, ..., Nach_8] // Von_8 ]
-     *
-     * @return mit BigDecimal.ZERO initialisierte Datenstruktur
-     */
-    private static BigDecimal[][] getEmptyDatastructure() {
-        final BigDecimal[][] datastructure = new BigDecimal[8][8];
-        Arrays.stream(datastructure).forEach(data -> Arrays.fill(data, BigDecimal.ZERO));
-        return datastructure;
-    }
-
-    private static BelastungsplanDataDTO getEmptyBelastungsplanData() {
-        final BelastungsplanDataDTO data = new BelastungsplanDataDTO();
-        fillEmptyBelastungsplanData(data);
-        data.setPercent(false);
-        data.setValues(getEmptyDatastructure());
-        return data;
-    }
-
-    private static BelastungsplanQjsDataDTO getEmptyBelastungsplanQjsData() {
-        final BelastungsplanQjsDataDTO data = new BelastungsplanQjsDataDTO();
-        fillEmptyBelastungsplanData(data);
-        data.setSumAll(BigDecimal.ZERO);
-        data.setValuesStrassenseite(new ArrayList<>());
-        data.setValuesVerkehrsbeziehungen(new ArrayList<>());
-        return data;
-    }
-
-    private static void fillEmptyBelastungsplanData(AbstractBelastungsplanDataDTO data) {
-        data.setLabel("");
-        data.setFilled(false);
     }
 
     private static boolean isVerkehrsbeziehungNachOrKreisverkehrSet(final Zeitintervall zeitintervall) {
@@ -395,160 +351,7 @@ public class ProcessZaehldatenBelastungsplanService {
                                         VALUE_TO_ROUND,
                                         options))));
 
-        if (zaehlung.getZaehlart().equals(Zaehlart.QJS.name())) {
-            final LadeBelastungsplanQjsDTO ladeBelastungsplan = new LadeBelastungsplanQjsDTO();
-            ladeBelastungsplan.setStreets(new String[8]);
-            return buildBelastungsplanQjsData(ladeBelastungsplan, options, zaehlung, ladeZaehldatumBelastungsplan);
-        } else {
-            final LadeBelastungsplanDTO ladeBelastungsplan = new LadeBelastungsplanDTO();
-            ladeBelastungsplan.setStreets(new String[8]);
-            return buildBelastungsplanDefaultData(ladeBelastungsplan, options, zaehlung, ladeZaehldatumBelastungsplan);
-        }
-    }
-
-    private LadeBelastungsplanQjsDTO buildBelastungsplanQjsData(final LadeBelastungsplanQjsDTO ladeBelastungsplan, final OptionsDTO options,
-            final Zaehlung zaehlung,
-            final Map<Verkehrsbeziehung, TupelTageswertZaehldatum> ladeZaehldatumBelastungsplan) {
-        ladeBelastungsplan.setValue1(getEmptyBelastungsplanQjsData());
-        ladeBelastungsplan.setValue2(getEmptyBelastungsplanQjsData());
-        ladeBelastungsplan.setValue3(getEmptyBelastungsplanQjsData());
-        final Map<Fahrzeug, AbstractBelastungsplanDataDTO> belastungsplanData = getBelastungsplanQjsData(ladeZaehldatumBelastungsplan, zaehlung);
-        zaehlung.getKnotenarme().forEach(knotenarm -> ladeBelastungsplan.getStreets()[knotenarm.getNummer() - 1] = knotenarm.getStrassenname());
-        if (options.getRadverkehr() && belastungsplanData.containsKey(Fahrzeug.RAD)) {
-            putFirstValueInBelastungsplan(ladeBelastungsplan, belastungsplanData, Fahrzeug.RAD);
-        } else if (options.getFussverkehr() && belastungsplanData.containsKey(Fahrzeug.FUSS)) {
-            putFirstValueInBelastungsplan(ladeBelastungsplan, belastungsplanData, Fahrzeug.FUSS);
-        }
-        markKIHochrechnung(zaehlung.getZaehldauer(), options.getZeitauswahl(), ladeBelastungsplan);
-        return ladeBelastungsplan;
-    }
-
-    private LadeBelastungsplanDTO buildBelastungsplanDefaultData(final LadeBelastungsplanDTO ladeBelastungsplan, final OptionsDTO options,
-            final Zaehlung zaehlung, final Map<Verkehrsbeziehung, TupelTageswertZaehldatum> ladeZaehldatumBelastungsplan) {
-        ladeBelastungsplan.setValue1(getEmptyBelastungsplanData());
-        ladeBelastungsplan.setValue2(getEmptyBelastungsplanData());
-        ladeBelastungsplan.setValue3(getEmptyBelastungsplanData());
-
-        final Map<Fahrzeug, AbstractBelastungsplanDataDTO> belastungsplanData = getBelastungsplanData(ladeZaehldatumBelastungsplan, zaehlung);
-        zaehlung.getKnotenarme().forEach(knotenarm -> ladeBelastungsplan.getStreets()[knotenarm.getNummer() - 1] = knotenarm.getStrassenname());
-        ladeBelastungsplan.setKreisverkehr(zaehlung.getKreisverkehr());
-
-        // Wenn KFZ, GV, SV, GV_Prozent oder SV_Prozent gesetzt ist, dürfen RAD und FUSS nicht angezeigt werden
-        if ((options.getGueterverkehrsanteilProzent() && belastungsplanData.containsKey(Fahrzeug.GV_P))
-                || (options.getSchwerverkehrsanteilProzent() && belastungsplanData.containsKey(Fahrzeug.SV_P))
-                || (options.getGueterverkehr() && belastungsplanData.containsKey(Fahrzeug.GV))
-                || (options.getSchwerverkehr() && belastungsplanData.containsKey(Fahrzeug.SV))
-                || (options.getKraftfahrzeugverkehr() && belastungsplanData.containsKey(Fahrzeug.KFZ))) {
-            if (options.getGueterverkehrsanteilProzent() && belastungsplanData.containsKey(Fahrzeug.GV_P)) {
-                putFirstValueInBelastungsplan(ladeBelastungsplan, belastungsplanData, Fahrzeug.GV_P);
-            }
-            if (options.getSchwerverkehrsanteilProzent() && belastungsplanData.containsKey(Fahrzeug.SV_P)) {
-                putFirstValueInBelastungsplan(ladeBelastungsplan, belastungsplanData, Fahrzeug.SV_P);
-            }
-            if (options.getGueterverkehr() && belastungsplanData.containsKey(Fahrzeug.GV)) {
-                putFirstValueInBelastungsplan(ladeBelastungsplan, belastungsplanData, Fahrzeug.GV);
-            }
-            if (options.getSchwerverkehr() && belastungsplanData.containsKey(Fahrzeug.SV)) {
-                putFirstValueInBelastungsplan(ladeBelastungsplan, belastungsplanData, Fahrzeug.SV);
-            }
-            if (options.getKraftfahrzeugverkehr() && belastungsplanData.containsKey(Fahrzeug.KFZ)) {
-                putFirstValueInBelastungsplan(ladeBelastungsplan, belastungsplanData, Fahrzeug.KFZ);
-            }
-        } else if (options.getRadverkehr() && belastungsplanData.containsKey(Fahrzeug.RAD)) {
-            putFirstValueInBelastungsplan(ladeBelastungsplan, belastungsplanData, Fahrzeug.RAD);
-        } else if (options.getFussverkehr() && belastungsplanData.containsKey(Fahrzeug.FUSS)) {
-            putFirstValueInBelastungsplan(ladeBelastungsplan, belastungsplanData, Fahrzeug.FUSS);
-        }
-
-        LadeBelastungsplanDTO ladeBelastungsplanSum = this.calculateSumsForLadeBelastungsplanDto(ladeBelastungsplan,
-                (BelastungsplanDataDTO) belastungsplanData.get(Fahrzeug.KFZ),
-                (BelastungsplanDataDTO) belastungsplanData.get(Fahrzeug.SV), (BelastungsplanDataDTO) belastungsplanData.get(Fahrzeug.GV));
-
-        markKIHochrechnung(zaehlung.getZaehldauer(), options.getZeitauswahl(), ladeBelastungsplanSum);
-
-        return ladeBelastungsplanSum;
-    }
-
-    private void markKIHochrechnung(final String zaehldauer, final String zeitauswahl,
-            final AbstractLadeBelastungsplanDTO<? extends AbstractBelastungsplanDataDTO> ladeBelastungsplanSum) {
-        // KI-Hochgerechnete Werte sollen im Belastungsplan entsprechend gekennzeichnet werden
-        if (Zeitauswahl.TAGESWERT.getCapitalizedName().equals(zeitauswahl) && List.of(Zaehldauer.DAUER_2_X_4_STUNDEN.toString(),
-                Zaehldauer.DAUER_13_STUNDEN.toString(), Zaehldauer.DAUER_16_STUNDEN.toString()).contains(zaehldauer)) {
-            Stream.of(
-                    ladeBelastungsplanSum.getValue1(),
-                    ladeBelastungsplanSum.getValue2(),
-                    ladeBelastungsplanSum.getValue3()).filter(v -> "RAD".equals(v.getLabel()))
-                    .forEach(v -> v.setLabel("RAD (KI-Hochrechnung)"));
-        }
-    }
-
-    @SuppressWarnings("unchecked")
-    private <T extends AbstractBelastungsplanDataDTO> void putFirstValueInBelastungsplan(AbstractLadeBelastungsplanDTO<T> ladeBelastungsplan,
-            Map<Fahrzeug, ? extends AbstractBelastungsplanDataDTO> belastungsplanData,
-            Fahrzeug value) {
-        ladeBelastungsplan.setValue3(ladeBelastungsplan.getValue2());
-        ladeBelastungsplan.setValue2(ladeBelastungsplan.getValue1());
-        ladeBelastungsplan.setValue1((T) belastungsplanData.get(value));
-    }
-
-    /**
-     * Reichert das übergebene LadeBelastungsplanDTO-Objekt um die Summen der einzelnen Knotenarme an.
-     *
-     * @param ladeBelastungsplan LadeBelastungsplanDTO-Objekt, welches um die Summen angereichert werden
-     *            soll
-     * @param dataKfz Datengrundlage von KFZ zur Berechnung der %-Anteile
-     * @param dataSv Datengrundlage von SV zur Berechnung der SV%-Anteile
-     * @param dataGv Datengrundlage von GV zur Berechnung der GV%-Anteile
-     * @return gibt das um alle Summen erweiterte LadeBelastungsplanDTO-Objekt zurück
-     */
-    private LadeBelastungsplanDTO calculateSumsForLadeBelastungsplanDto(final LadeBelastungsplanDTO ladeBelastungsplan, final BelastungsplanDataDTO dataKfz,
-            final BelastungsplanDataDTO dataSv, final BelastungsplanDataDTO dataGv) {
-
-        Map<String, BigDecimal[]> sumsKfz = null;
-        Map<String, BigDecimal[]> sumsSv = null;
-        Map<String, BigDecimal[]> sumsGv = null;
-
-        if (dataKfz != null && dataKfz.isFilled()) {
-            if (ladeBelastungsplan.isKreisverkehr()) {
-                sumsKfz = BelastungsplanCalculator.calcSumsKreisverkehr(dataKfz.getValues());
-            } else {
-                sumsKfz = BelastungsplanCalculator.calcSumsKreuzung(dataKfz.getValues());
-            }
-        }
-        if (dataSv != null && dataSv.isFilled()) {
-            if (ladeBelastungsplan.isKreisverkehr()) {
-                sumsSv = BelastungsplanCalculator.calcSumsKreisverkehr(dataSv.getValues());
-            } else {
-                sumsSv = BelastungsplanCalculator.calcSumsKreuzung(dataSv.getValues());
-            }
-        }
-        if (dataGv != null && dataGv.isFilled()) {
-            if (ladeBelastungsplan.isKreisverkehr()) {
-                sumsGv = BelastungsplanCalculator.calcSumsKreisverkehr(dataGv.getValues());
-            } else {
-                sumsGv = BelastungsplanCalculator.calcSumsKreuzung(dataGv.getValues());
-            }
-        }
-
-        if (ladeBelastungsplan.getValue1().isFilled()) {
-            ladeBelastungsplan.setValue1(
-                    BelastungsplanCalculator.calculateSumsForBelastungsplanDataDto(ladeBelastungsplan.getValue1(), sumsKfz, sumsSv, sumsGv,
-                            ladeBelastungsplan.isKreisverkehr()));
-        }
-
-        if (ladeBelastungsplan.getValue2().isFilled()) {
-            ladeBelastungsplan.setValue2(
-                    BelastungsplanCalculator.calculateSumsForBelastungsplanDataDto(ladeBelastungsplan.getValue2(), sumsKfz, sumsSv, sumsGv,
-                            ladeBelastungsplan.isKreisverkehr()));
-        }
-
-        if (ladeBelastungsplan.getValue3().isFilled()) {
-            ladeBelastungsplan.setValue3(
-                    BelastungsplanCalculator.calculateSumsForBelastungsplanDataDto(ladeBelastungsplan.getValue3(), sumsKfz, sumsSv, sumsGv,
-                            ladeBelastungsplan.isKreisverkehr()));
-        }
-
-        return ladeBelastungsplan;
+        return belastungsplanDataServiceFactory.getBelastungsplanDataService(zaehlung).buildBelastungsplanData(options, zaehlung, ladeZaehldatumBelastungsplan);
     }
 
     /**
@@ -677,144 +480,6 @@ public class ProcessZaehldatenBelastungsplanService {
     }
 
     /**
-     * Liefert eine {@link BelastungsplanDataDTO} pro Fahrzeugklasse mit den Daten für den
-     * Belastungsplan
-     *
-     * @param zaehldatenJeVerkehrsbeziehung aus der DB ermittelten Werte
-     * @param zaehlung wird benötigt zur überprüfung, ob welche Fahrzeug gezählt wurden
-     * @return eine Map mit Key: Fahrzeug und Value:BelastungsplanDataDTO.
-     */
-    public Map<Fahrzeug, AbstractBelastungsplanDataDTO> getBelastungsplanData(
-            final Map<Verkehrsbeziehung, TupelTageswertZaehldatum> zaehldatenJeVerkehrsbeziehung,
-            final Zaehlung zaehlung) {
-        final Map<Fahrzeug, AbstractBelastungsplanDataDTO> returnValue = new HashMap<>();
-
-        final BelastungsplanDataDTO belastungsplanDataKfz = new BelastungsplanDataDTO();
-        belastungsplanDataKfz.setFilled(zaehlung.getKategorien().contains(Fahrzeug.KFZ));
-        belastungsplanDataKfz.setPercent(false);
-        belastungsplanDataKfz.setLabel(Fahrzeug.KFZ.getName());
-        belastungsplanDataKfz.setValues(getEmptyDatastructure());
-
-        final BelastungsplanDataDTO belastungsplanDataSv = new BelastungsplanDataDTO();
-        belastungsplanDataSv.setFilled(zaehlung.getKategorien().contains(Fahrzeug.SV));
-        belastungsplanDataSv.setPercent(false);
-        belastungsplanDataSv.setLabel(Fahrzeug.SV.getName());
-        belastungsplanDataSv.setValues(getEmptyDatastructure());
-
-        final BelastungsplanDataDTO belastungsplanDataGv = new BelastungsplanDataDTO();
-        belastungsplanDataGv.setFilled(zaehlung.getKategorien().contains(Fahrzeug.GV));
-        belastungsplanDataGv.setPercent(false);
-        belastungsplanDataGv.setLabel(Fahrzeug.GV.getName());
-        belastungsplanDataGv.setValues(getEmptyDatastructure());
-
-        final BelastungsplanDataDTO belastungsplanDataRad = new BelastungsplanDataDTO();
-        belastungsplanDataRad.setFilled(zaehlung.getKategorien().contains(Fahrzeug.RAD));
-        belastungsplanDataRad.setPercent(false);
-        belastungsplanDataRad.setLabel(Fahrzeug.RAD.getName());
-        belastungsplanDataRad.setValues(getEmptyDatastructure());
-
-        final BelastungsplanDataDTO belastungsplanDataFuss = new BelastungsplanDataDTO();
-        belastungsplanDataFuss.setFilled(zaehlung.getKategorien().contains(Fahrzeug.FUSS));
-        belastungsplanDataFuss.setPercent(false);
-        belastungsplanDataFuss.setLabel(Fahrzeug.FUSS.getName());
-        belastungsplanDataFuss.setValues(getEmptyDatastructure());
-
-        final BelastungsplanDataDTO belastungsplanDataSvProzent = new BelastungsplanDataDTO();
-        belastungsplanDataSvProzent.setFilled(belastungsplanDataKfz.isFilled() && belastungsplanDataSv.isFilled());
-        belastungsplanDataSvProzent.setPercent(true);
-        belastungsplanDataSvProzent.setLabel(Fahrzeug.SV_P.getName());
-        belastungsplanDataSvProzent.setValues(getEmptyDatastructure());
-
-        final BelastungsplanDataDTO belastungsplanDataGvProzent = new BelastungsplanDataDTO();
-        belastungsplanDataGvProzent.setFilled(belastungsplanDataKfz.isFilled() && belastungsplanDataGv.isFilled());
-        belastungsplanDataGvProzent.setPercent(true);
-        belastungsplanDataGvProzent.setLabel(Fahrzeug.GV_P.getName());
-        belastungsplanDataGvProzent.setValues(getEmptyDatastructure());
-
-        zaehldatenJeVerkehrsbeziehung.forEach((verkehrsbeziehung, tupelTageswertZaehldatum) -> {
-            final int index1;
-            final int index2;
-            if (isKreisverkehr(verkehrsbeziehung)) {
-                // Von-Knotennummer - 1
-                index1 = verkehrsbeziehung.getVon() - 1;
-                // HINEIN = 0, VORBEI = 1, HERAUS = 2
-                index2 = switch (verkehrsbeziehung.getFahrbewegungKreisverkehr()) {
-                case HINEIN -> 0;
-                case VORBEI -> 1;
-                case HERAUS -> 2;
-                };
-            } else {
-                index1 = verkehrsbeziehung.getVon() - 1;
-                index2 = verkehrsbeziehung.getNach() - 1;
-            }
-
-            belastungsplanDataKfz.getValues()[index1][index2] = tupelTageswertZaehldatum.getLadeZaehldatum().getKfz();
-            belastungsplanDataSv.getValues()[index1][index2] = tupelTageswertZaehldatum.getLadeZaehldatum().getSchwerverkehr();
-            belastungsplanDataGv.getValues()[index1][index2] = tupelTageswertZaehldatum.getLadeZaehldatum().getGueterverkehr();
-
-            if (belastungsplanDataSvProzent.isFilled()) {
-                belastungsplanDataSvProzent.getValues()[index1][index2] = CalculationUtil
-                        .calculateAnteilProzent(belastungsplanDataSv.getValues()[index1][index2], belastungsplanDataKfz.getValues()[index1][index2]);
-            }
-            if (belastungsplanDataGvProzent.isFilled()) {
-                belastungsplanDataGvProzent.getValues()[index1][index2] = CalculationUtil
-                        .calculateAnteilProzent(belastungsplanDataGv.getValues()[index1][index2], belastungsplanDataKfz.getValues()[index1][index2]);
-            }
-
-            belastungsplanDataRad.getValues()[index1][index2] = BigDecimal.valueOf(
-                    Objects.requireNonNullElse(
-                            tupelTageswertZaehldatum.getLadeZaehldatum().getFahrradfahrer(),
-                            0));
-
-            if (!tupelTageswertZaehldatum.getIsTageswert()) {
-                belastungsplanDataFuss.getValues()[index1][index2] = BigDecimal.valueOf(
-                        Objects.requireNonNullElse(
-                                tupelTageswertZaehldatum.getLadeZaehldatum().getFussgaenger(),
-                                0));
-            }
-        });
-
-        if (belastungsplanDataKfz.isFilled()) {
-            returnValue.put(Fahrzeug.KFZ, belastungsplanDataKfz);
-        }
-        if (belastungsplanDataSv.isFilled()) {
-            returnValue.put(Fahrzeug.SV, belastungsplanDataSv);
-        }
-        if (belastungsplanDataGv.isFilled()) {
-            returnValue.put(Fahrzeug.GV, belastungsplanDataGv);
-        }
-        if (belastungsplanDataRad.isFilled()) {
-            returnValue.put(Fahrzeug.RAD, belastungsplanDataRad);
-        }
-        if (belastungsplanDataFuss.isFilled()) {
-            returnValue.put(Fahrzeug.FUSS, belastungsplanDataFuss);
-        }
-        if (belastungsplanDataSvProzent.isFilled()) {
-            returnValue.put(Fahrzeug.SV_P, belastungsplanDataSvProzent);
-        }
-        if (belastungsplanDataGvProzent.isFilled()) {
-            returnValue.put(Fahrzeug.GV_P, belastungsplanDataGvProzent);
-        }
-        return returnValue;
-    }
-
-    public Map<Fahrzeug, AbstractBelastungsplanDataDTO> getBelastungsplanQjsData(
-            final Map<Verkehrsbeziehung, TupelTageswertZaehldatum> zaehldatenJeVerkehrsbeziehung,
-            final Zaehlung zaehlung) {
-        final Map<Fahrzeug, AbstractBelastungsplanDataDTO> returnValue = new HashMap<>();
-
-        if (zaehlung.getKategorien().contains(Fahrzeug.RAD)) {
-            returnValue.put(Fahrzeug.RAD,
-                    buildBelastungsplanQjsDataForFahrzeug(Fahrzeug.RAD, LadeZaehldatumDTO::getFahrradfahrer, zaehldatenJeVerkehrsbeziehung));
-        }
-        if (zaehlung.getKategorien().contains(Fahrzeug.FUSS)) {
-            returnValue.put(Fahrzeug.FUSS,
-                    buildBelastungsplanQjsDataForFahrzeug(Fahrzeug.FUSS, LadeZaehldatumDTO::getFussgaenger, zaehldatenJeVerkehrsbeziehung));
-        }
-        return returnValue;
-    }
-
-    /**
      * Auf Basis der Zaehlungs-Id im Parameter wird die Zaehlung aus der DB extrahiert.
      *
      * @param zaehlungId Die Zaehlungs-ID für die {@link Zaehlung}.
@@ -834,53 +499,6 @@ public class ProcessZaehldatenBelastungsplanService {
             throw new DataNotFoundException("Die Zählung " + zaehlungId + " wurde nicht gefunden");
         }
         return zaehlungOptional.get();
-    }
-
-    private BelastungsplanQjsDataDTO buildBelastungsplanQjsDataForFahrzeug(
-            final Fahrzeug fz,
-            final Function<LadeZaehldatumDTO, Integer> reader,
-            final Map<Verkehrsbeziehung, TupelTageswertZaehldatum> zaehldatenJeVerkehrsbeziehung) {
-        final BelastungsplanQjsDataDTO belastungsplanData = getEmptyBelastungsplanQjsData();
-        belastungsplanData.setFilled(true);
-        belastungsplanData.setLabel(fz.getName());
-        belastungsplanData.setSumAll(BigDecimal.ZERO);
-        belastungsplanData.setValuesStrassenseite(new ArrayList<>());
-        belastungsplanData.setValuesVerkehrsbeziehungen(new ArrayList<>());
-        zaehldatenJeVerkehrsbeziehung.forEach((verkehrsbeziehung, tupelTageswertZaehldatum) -> {
-            checkVerkehrsbeziehungenForDuplicates(belastungsplanData.getValuesVerkehrsbeziehungen(), verkehrsbeziehung);
-            var value = new BelastungsplanQjsDataDTO.VerkehrsbeziehungValue(verkehrsbeziehung.getVon(), verkehrsbeziehung.getNach(),
-                    verkehrsbeziehung.getStrassenseite(),
-                    BigDecimal.valueOf(Objects.requireNonNullElse(reader.apply(tupelTageswertZaehldatum.getLadeZaehldatum()), 0)));
-            belastungsplanData.getValuesVerkehrsbeziehungen().add(value);
-
-            Optional<BelastungsplanQjsDataDTO.StrassenseiteValue> valueStrassenseite = belastungsplanData.getValuesStrassenseite().stream()
-                    .filter(bez -> bez.getStrassenseite() == verkehrsbeziehung.getStrassenseite()).findFirst();
-            if (valueStrassenseite.isPresent()) {
-                BigDecimal oldValue = valueStrassenseite.get().getValue();
-                BigDecimal newValue = oldValue
-                        .add(BigDecimal.valueOf(Objects.requireNonNullElse(reader.apply(tupelTageswertZaehldatum.getLadeZaehldatum()), 0)));
-                valueStrassenseite.get().setValue(newValue);
-            } else {
-                BelastungsplanQjsDataDTO.StrassenseiteValue valueStrassenseite2 = new BelastungsplanQjsDataDTO.StrassenseiteValue(
-                        verkehrsbeziehung.getStrassenseite());
-                valueStrassenseite2.setValue(BigDecimal.valueOf(Objects.requireNonNullElse(reader.apply(tupelTageswertZaehldatum.getLadeZaehldatum()), 0)));
-                belastungsplanData.getValuesStrassenseite().add(valueStrassenseite2);
-            }
-            belastungsplanData
-                    .setSumAll(belastungsplanData.getSumAll()
-                            .add(BigDecimal.valueOf(Objects.requireNonNullElse(reader.apply(tupelTageswertZaehldatum.getLadeZaehldatum()), 0))));
-        });
-        return belastungsplanData;
-    }
-
-    private void checkVerkehrsbeziehungenForDuplicates(
-            List<BelastungsplanQjsDataDTO.VerkehrsbeziehungValue> valuesVerkehrsbeziehungen,
-            Verkehrsbeziehung verkehrsbeziehung) {
-        if (valuesVerkehrsbeziehungen.stream().anyMatch(bez -> (bez.getVon() == verkehrsbeziehung.getVon())
-                && (bez.getNach() == verkehrsbeziehung.getNach()) && (bez.getStrassenseite() == verkehrsbeziehung.getStrassenseite()))) {
-            log.error("Fehler beim Berechnen der Daten: doppelte Verkehrsbeziehungen");
-            throw new IllegalStateException("Fehler beim Berechnen der Daten");
-        }
     }
 
     @AllArgsConstructor
